Move c11y toggle into `PlankHeading`s. More polish. Fix missing translations.
