+++ packages/ui/lit-grid/src/dx-grid.pcss
@@ -14,8 +14,8 @@ dx-grid {
   grid-template-rows: min-content 1fr min-content;
   font-variant-numeric: tabular-nums;
 
-  min-inline-size: 4rem;
-  min-block-size: 4rem;
+  min-inline-size: 64px;
+  min-block-size: 30px;
   inline-size: 100%;
   block-size: 100%;
   max-inline-size: 100dvw;
