+++ packages/ui/react-ui-stack/src/next/index.ts
@@ -5,4 +5,3 @@
 export * from './Stack';
 export * from './StackItem';
 export * from './PlankHeading';
-export { default as translations } from './translations';
