--- packages/ui/react-ui-stack/src/next/translations.ts
@@ -1,23 +1,0 @@
-//
-// Copyright 2024 DXOS.org
-//
-
-export const translationKey = 'react-ui-stack';
-
-export default [
-  {
-    'en-US': {
-      [translationKey]: {
-        'drag handle label': 'Press space or enter to begin dragging this item.',
-        'resize section label': 'Drag to resize section',
-        'pin start label': 'Pin to the left sidebar',
-        'pin end label': 'Pin to the right sidebar',
-        'increment start label': 'Move to the left',
-        'increment end label': 'Move to the right',
-        'solo plank label': 'Toggle solo mode',
-        'close label': 'Close',
-        'minify label': 'Minify',
-      },
-    },
-  },
-];
