+++ packages/ui/react-ui-stack/src/translations.ts
@@ -16,6 +16,15 @@ export default [
         'add section after label': 'Add after',
         'expand label': 'Expand',
         'collapse label': 'Collapse',
+        'drag handle label': 'Press space or enter to begin dragging this item.',
+        'resize label': 'Drag to resize',
+        'pin start label': 'Pin to the left sidebar',
+        'pin end label': 'Pin to the right sidebar',
+        'increment start label': 'Move to the left',
+        'increment end label': 'Move to the right',
+        'solo plank label': 'Toggle solo mode',
+        'close label': 'Close',
+        'minify label': 'Minify',
       },
     },
   },
